+++ apps/box/src/inngest/functions/index.ts
@@ -1,6 +1,6 @@
 import { refreshToken } from './token/refresh-token';
-import { synchronizeUsers } from './users/synchronize-users';
-import { scheduleUsersSynchronize } from './users/schedule-users-synchronize';
+import { synchronizeUsers } from './users/sync-users';
+import { scheduleUsersSynchronize } from './users/schedule-users-sync';
 import { deleteUser } from './users/delete-user';
 import { removeOrganisation } from './organisations/remove-organisation';
