+++ apps/box/src/inngest/functions/users/schedule-users-sync.test.ts
@@ -2,7 +2,7 @@ import { expect, test, describe, beforeAll, vi, afterAll } from 'vitest';
 import { createInngestFunctionMock } from '@elba-security/test-utils';
 import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
-import { scheduleUsersSynchronize } from './schedule-users-synchronize';
+import { scheduleUsersSynchronize } from './schedule-users-sync';
 
 const now = Date.now();
