+++ apps/box/src/inngest/functions/users/sync-users.test.ts
@@ -5,7 +5,7 @@ import * as usersConnector from '@/connectors/box/users';
 import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
 import { encrypt } from '@/common/crypto';
-import { synchronizeUsers } from './synchronize-users';
+import { synchronizeUsers } from './sync-users';
 
 const organisation = {
   id: '45a76301-f1dd-4a77-b12f-9d7d3fca3c90',
